Fix AABB placement and handling of multiple boxes.
A new box can of course overlap with an arbitrary number of old ones, so
even for simple cases you need to handle multiple boxes properly. (And,
in fact, even the case of overlapping with just one wasn't handled
properly before this change.)

Decomposition becomes a lot harder due to this and can't be done the
semi-hard-coded way it was done, so that dead code was removed. Consider
for instance the following 2D situation:

   +-------+
----+  new |
old1|  box |
----+      |
------+    |
 old2 |----+
      |
+-----+

Decomposing the new box in the above results in more than the 9 initial
subboxes that one might expect.

Of course, the old code could still be used by recursively decomposing
one box at a time: in the above, first new with old1, then each result
of that with old2. It just sounds horribly inefficient in the worst
case.

The generic way to do it would be taking the difference of the convex
polygons followed by tessellating the resulting polygon into AABBs. Not
necessarily very trivial.

Note that the removed dead code includes the utility operations
fuseWith and getOverlapWith, which could be useful in the future.