Set TRDS's earliest time lazily when it is loaded.
This allows us to postpone copying of the initial space until TRDS is
loaded, meaning we can avoid doing it completely if TRDS is never used.
This can be a noticeable speedup for some programs.

Before, we could just point the IP at (0,0) and start over when jumping.
Now, we actually have to save a snapshot of the interpreter state. Hence
the FungeState struct in ccbi.fungestate, which contains everything we
want to save.

Something notable which could be saved but wasn't added to FungeState is
TURT state, mostly due to simply not doing it being the simplest way of
avoiding circular dependency issues.

Most of the changes are due to moving stuff like the current Funge-Space
and the ips array to FungeState. The relevant changes are in TRDS,
FungeMachine, and FungeState.