Fix and test subsumeArea, fix canDirectCopy usage.
subsumeArea worked properly only when called from subsume. Modified
fusion-overlapping.b98.t to trigger the now-fixed bug. Not 100% sure
about the 3D fix and modifying the .t98.t wasn't as trivial as for the
.b98.t so that case remains incompletely tested.

In addition, canDirectCopy's size test wasn't always looking at the
right size, so fixed that too.