Wrap and skip markers more cleverly and correctly.
Consolidates the last couple of anticipatory changes: this is what they
were made for.

The cleverness should translate to speed in cases where there's hundreds
or thousands of unallocated spaces to skip over. Other than that it
should hopefully be faster, but it might be a bit slower for the trivial
cases like just a few spaces in an allocated box. Testing shows that
Mycology speed does go up a bit, from 0.03-0.04 s to a fairly constant
0.04 s.

The correctness means that # over file edge now only skips an edge
column if it was at the edge of addressable space.